Enum sourcemap::Error

pub enum Error {
    Io(Error),
    Scroll(Error),
    Utf8(Utf8Error),
    BadJson(Error),
    VlqLeftover,
    VlqNoValues,
    VlqOverflow,
    BadSegmentSize(u32),
    BadSourceReference(u32),
    BadNameReference(u32),
    IndexedSourcemap,
    RegularSourcemap,
    InvalidDataUrl,
    CannotFlatten(String),
    InvalidRamBundleMagic,
    InvalidRamBundleIndex,
    InvalidRamBundleEntry,
    NotARamBundle,
}

Represents different failure cases

Variants
Io(Error)

a std::io error

Scroll(Error)

a scroll error

Utf8(Utf8Error)

a std::str::Utf8Error

BadJson(Error)

a JSON parsing related failure

VlqLeftover

a VLQ string was malformed and data was left over

VlqNoValues

a VLQ string was empty and no values could be decoded.

VlqOverflow

Overflow in Vlq handling

BadSegmentSize(u32)

a mapping segment had an unsupported size

BadSourceReference(u32)

a reference to a non existing source was encountered

BadNameReference(u32)

a reference to a non existing name was encountered

IndexedSourcemap

Indicates that an indexed sourcemap was encountered when a regular sourcemap was expected

RegularSourcemap

Indicates that an regular (non-indexed) sourcemap was when a sourcemap index was expected

InvalidDataUrl

Indicates an invalid data URL

CannotFlatten(String)

Flatten failed

InvalidRamBundleMagic

The magic of a RAM bundle did not match

InvalidRamBundleIndex

The RAM bundle index was malformed

InvalidRamBundleEntry

A RAM bundle entry was invalid

NotARamBundle

Tried to operate on a non RAM bundle file