Function sourcemap::ram_bundle::split_ram_bundle

pub fn split_ram_bundle<'a, 'b>(
    ram_bundle: &'b RamBundle<'a>, 
    smi: &SourceMapIndex
) -> Result<SplitRamBundleModuleIter<'a, 'b>>

Decontstructs a RAM bundle into a sequence of sources and their sourcemaps

With the help of the RAM bundle's indexed sourcemap, the bundle is split into modules, where each module is represented by its minified source and the corresponding sourcemap that we recover from the initial indexed sourcemap.