Struct sourcemap::RawToken

pub struct RawToken {
    pub dst_line: u32,
    pub dst_col: u32,
    pub src_line: u32,
    pub src_col: u32,
    pub src_id: u32,
    pub name_id: u32,
}

Represents a raw token

Raw tokens are used internally to represent the sourcemap in a memory efficient way. If you construct sourcemaps yourself then you need to create these objects, otherwise they are invisible to you as a user.

Fields

dst_line

the destination (minified) line number

dst_col

the destination (minified) column number

src_line

the source line number

src_col

the source line column

src_id

source identifier

name_id

name identifier (!0 in case there is no associated name)