sourcemap
Struct SourceMapHermes

pub struct SourceMapHermes { /* private fields */ }

Represents a react-native-style SourceMap, which has additional scope information embedded.

Implementations

impl SourceMapHermes

pub fn from_reader<R: Read>(rdr: R) -> Result<Self>

Creates a sourcemap from a reader over a JSON stream in UTF-8 format.

See SourceMap::from_reader

pub fn from_slice(slice: &[u8]) -> Result<Self>

Creates a sourcemap from a reader over a JSON byte slice in UTF-8 format.

See SourceMap::from_slice

pub fn to_writer<W: Write>(&self, w: W) -> Result<()>

Writes a sourcemap into a writer.

See SourceMap::to_writer

pub fn get_original_function_name(&self, bytecode_offset: u32) -> Option<&str>

Given a bytecode offset, this will find the enclosing scopes function name.

pub fn get_scope_for_token(&self, token: Token<'_>) -> Option<&str>

Resolves the name of the enclosing function for the given Token.

pub fn rewrite(self, options: &RewriteOptions<'_>) -> Result<Self>

This rewrites the sourcemap according to the provided rewrite options.

See SourceMap::rewrite