Enum sourcemap::Error

pub enum Error {
    Io(Error),
    Utf8(Utf8Error),
    BadJson(usize, usize, String),
    VlqLeftover,
    VlqNoValues,
    BadSegmentSize(u32),
    BadSourceReference(u32),
    BadNameReference(u32),
    IndexedSourcemap,
    RegularSourcemap,
    InvalidDataUrl,
}

Represents different failure cases.

Variants

Io

a std::io error

Utf8

a std::str::Utf8Error

BadJson

a JSON parsing related failure

VlqLeftover

a VLQ string was malformed and data was left over

VlqNoValues

a VLQ string was empty and no values could be decoded.

BadSegmentSize

a mapping segment had an unsupported size

BadSourceReference

a reference to a non existing source was encountered

BadNameReference

a reference to a non existing name was encountered

IndexedSourcemap

Indicates that an indexed sourcemap was encountered when a regular sourcemap was expected

RegularSourcemap

Indicates that an regular (non-indexed) sourcemap was when a sourcemap index was expected

InvalidDataUrl

Indicates an invalid data URL

Methods

impl Error

fn source_location(&self) -> Option<(usize, usize)>

If the error originated in a JSON file that can be located, then this method returns that location.